vfio: harden RTL8168-msix-quirk against exploits
The VFIO RTL8168 quirk redirects a hw backdoor in BAR2 at 0x70-0x77
(which provides access to the MSI-X table) from to real hardware to
the emulated MSIX-table.
In order to harden the redirection code against malicious guest
activities, a few counter measures against potential leakage etc.
are added with this commit.

Signed-off-by: Thorsten Kohfeldt <thorsten.kohfeldt@gmx.de>